refactor: 이미지 파일명 및 alt 속성 수정
