+++ components/common/LoadingLogo.tsx
@@ -1,6 +1,7 @@
 import styled from '@emotion/styled';
 import Image from 'next/image';
-import { loading, moon } from 'public/icons';
+import loading from 'public/icons/loading.gif';
+import loadingBackground from 'public/icons/loading-background.png';
 
 const LoadingLogo = () => {
     return (
@@ -10,7 +11,7 @@ const LoadingLogo = () => {
                     <Image src={loading} alt="토끼가 카드를 고르는 중" />
                 </span>
                 <span>
-                    <Image src={moon} alt="orange circle background" />
+                    <Image src={loadingBackground} alt="토끼 오렌지색 배경" />
                 </span>
             </StyledImages>
         </>
