+++ components/common/Logo.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import logo_light from 'public/icons/main-logo_light.png';
+import logo_light from 'public/icons/main-logo-light.png';
 
 const Logo = () => {
     return <Image src={logo_light} layout="fixed" alt="달토끼" />;
