--- public/icons/index.ts
@@ -1,2 +1,0 @@
-export { default as loading } from './loading.gif';
-export { default as moon } from './moon.png';
